requirements.txt
fastapi
uvicorn
Pillow
pdf2image
python-magic
openai
